requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
pillow==11.1.0
pydantic==2.10.4
pydantic_core==2.27.2
python-multipart==0.0.20
setuptools==75.7.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
uvicorn==0.34.0
